Rendering process should support async operations
Fixes #54
