+++ examples/fractals/output.pdf
@@ -3,7 +3,7 @@
 1 0 obj
 <<
 /Keywords ()
-/CreationDate (D:20170421142455)
+/CreationDate (D:20170422181548)
 /Creator (react-pdf)
 /Producer (react-pdf)
 >>
@@ -1677,152 +1677,159 @@ ET
 endstream
 endobj
 xref
-0 139
+0 146
 0000000000 65535 f
-0000009331 00000 n
+0000011178 00000 n
 0000000015 00000 n
-0000010768 00000 n
-0000000144 00000 n
-0000000095 00000 n
-0000010580 00000 n
-0000002076 00000 n
-0000002010 00000 n
-0000002249 00000 n
-0000011102 00000 n
-0000001877 00000 n
-0000001944 00000 n
-0000002585 00000 n
-0000002394 00000 n
-0000011436 00000 n
-0000003064 00000 n
-0000002919 00000 n
-0000002729 00000 n
-0000001810 00000 n
-0000003252 00000 n
-0000011770 00000 n
-0000003729 00000 n
-0000001743 00000 n
-0000003586 00000 n
-0000003398 00000 n
-0000001676 00000 n
-0000003916 00000 n
-0000012103 00000 n
-0000001475 00000 n
-0000001609 00000 n
-0000004249 00000 n
-0000004061 00000 n
-0000001542 00000 n
-0000004580 00000 n
-0000004392 00000 n
-0000012249 00000 n
-0000005397 00000 n
-0000004917 00000 n
-0000004725 00000 n
-0000001408 00000 n
-0000005252 00000 n
-0000005063 00000 n
-0000001341 00000 n
-0000005585 00000 n
-0000013111 00000 n
-0000006400 00000 n
-0000001274 00000 n
-0000005920 00000 n
-0000005732 00000 n
-0000001207 00000 n
-0000006253 00000 n
-0000006066 00000 n
-0000001140 00000 n
-0000006590 00000 n
-0000008728 00000 n
-0000000872 00000 n
-0000001073 00000 n
-0000006926 00000 n
-0000006736 00000 n
-0000001006 00000 n
-0000007260 00000 n
-0000007071 00000 n
-0000000939 00000 n
-0000007594 00000 n
-0000007407 00000 n
-0000013929 00000 n
-0000009917 00000 n
-0000007929 00000 n
-0000007740 00000 n
-0000010105 00000 n
-0000009465 00000 n
-0000009599 00000 n
-0000000805 00000 n
-0000009771 00000 n
-0000000218 00000 n
-0000000742 00000 n
-0000009532 00000 n
-0000015474 00000 n
-0000011246 00000 n
-0000009197 00000 n
-0000009398 00000 n
-0000010437 00000 n
-0000010913 00000 n
-0000010249 00000 n
-0000009264 00000 n
-0000015270 00000 n
-0000012440 00000 n
-0000008996 00000 n
-0000009130 00000 n
-0000011581 00000 n
-0000009063 00000 n
-0000011915 00000 n
-0000016653 00000 n
-0000012921 00000 n
-0000008929 00000 n
-0000012775 00000 n
-0000008862 00000 n
-0000012586 00000 n
-0000016990 00000 n
-0000013594 00000 n
-0000008795 00000 n
-0000013782 00000 n
-0000013448 00000 n
-0000013258 00000 n
-0000017327 00000 n
-0000008528 00000 n
-0000008660 00000 n
-0000008074 00000 n
-0000014118 00000 n
-0000008592 00000 n
-0000017662 00000 n
-0000015793 00000 n
-0000015982 00000 n
-0000015406 00000 n
-0000015647 00000 n
-0000015338 00000 n
-0000017998 00000 n
-0000016462 00000 n
-0000015202 00000 n
-0000016318 00000 n
-0000016128 00000 n
-0000014862 00000 n
-0000015066 00000 n
-0000015134 00000 n
-0000016799 00000 n
-0000018143 00000 n
-0000017473 00000 n
-0000014998 00000 n
-0000017137 00000 n
-0000014930 00000 n
-0000017808 00000 n
-0000018333 00000 n
-0000014794 00000 n
-0000018668 00000 n
-0000018478 00000 n
-0000014726 00000 n
-0000014265 00000 n
-0000014662 00000 n
+0000010990 00000 n
+0000000205 00000 n
+0000000125 00000 n
+0000011512 00000 n
+0000002509 00000 n
+0000000254 00000 n
+0000002270 00000 n
+0000009607 00000 n
+0000002654 00000 n
+0000002336 00000 n
+0000002204 00000 n
+0000002845 00000 n
+0000011656 00000 n
+0000003512 00000 n
+0000002137 00000 n
+0000003179 00000 n
+0000002989 00000 n
+0000002070 00000 n
+0000011991 00000 n
+0000004176 00000 n
+0000003324 00000 n
+0000002003 00000 n
+0000003846 00000 n
+0000003658 00000 n
+0000001936 00000 n
+0000009406 00000 n
+0000004652 00000 n
+0000003989 00000 n
+0000001869 00000 n
+0000004509 00000 n
+0000004321 00000 n
+0000001802 00000 n
+0000004840 00000 n
+0000013185 00000 n
+0000005845 00000 n
+0000001735 00000 n
+0000005177 00000 n
+0000004985 00000 n
+0000001668 00000 n
+0000005512 00000 n
+0000005323 00000 n
+0000001601 00000 n
+0000013331 00000 n
+0000006850 00000 n
+0000005657 00000 n
+0000001534 00000 n
+0000006180 00000 n
+0000005992 00000 n
+0000001467 00000 n
+0000006513 00000 n
+0000006326 00000 n
+0000001400 00000 n
+0000014192 00000 n
+0000007667 00000 n
+0000006660 00000 n
+0000001333 00000 n
+0000007186 00000 n
+0000006996 00000 n
+0000001266 00000 n
+0000007520 00000 n
+0000007331 00000 n
+0000001199 00000 n
+0000007854 00000 n
+0000008334 00000 n
+0000009875 00000 n
+0000001132 00000 n
+0000008189 00000 n
+0000008000 00000 n
+0000001065 00000 n
+0000000328 00000 n
+0000010009 00000 n
+0000010181 00000 n
+0000001002 00000 n
+0000009942 00000 n
+0000000920 00000 n
+0000015900 00000 n
+0000011323 00000 n
+0000009674 00000 n
+0000010515 00000 n
+0000010327 00000 n
+0000009741 00000 n
+0000009808 00000 n
+0000010847 00000 n
+0000010659 00000 n
+0000016880 00000 n
+0000012325 00000 n
+0000011846 00000 n
+0000009540 00000 n
+0000012513 00000 n
+0000012180 00000 n
+0000009473 00000 n
+0000017024 00000 n
+0000013521 00000 n
+0000009272 00000 n
+0000012850 00000 n
+0000012659 00000 n
+0000009339 00000 n
+0000012996 00000 n
+0000017361 00000 n
+0000014004 00000 n
+0000009205 00000 n
+0000013858 00000 n
+0000013668 00000 n
+0000009138 00000 n
+0000017699 00000 n
+0000008938 00000 n
+0000009070 00000 n
+0000014528 00000 n
+0000014339 00000 n
+0000009002 00000 n
+0000018035 00000 n
+0000016544 00000 n
+0000008856 00000 n
+0000016036 00000 n
+0000015968 00000 n
+0000016209 00000 n
+0000018370 00000 n
+0000017215 00000 n
+0000015764 00000 n
+0000016355 00000 n
+0000016690 00000 n
+0000015832 00000 n
+0000018705 00000 n
+0000017889 00000 n
+0000015628 00000 n
+0000015696 00000 n
+0000017552 00000 n
+0000019040 00000 n
+0000015492 00000 n
+0000018224 00000 n
+0000015560 00000 n
+0000014675 00000 n
+0000018895 00000 n
+0000015424 00000 n
+0000018560 00000 n
+0000015356 00000 n
+0000019230 00000 n
+0000015288 00000 n
+0000015224 00000 n
+0000015140 00000 n
+0000000852 00000 n
+0000008788 00000 n
+0000015072 00000 n
 trailer
 <<
-/Size 138
+/Size 145
 /Root 3 0 R
 /Info 1 0 R
 >>
 startxref
-18814
+19376
 %%EOF
