+++ packages/react-pdf/src/elements/GraphicState.js
@@ -3,9 +3,13 @@ import { pdfObject, pdfStream } from '../utils/pdf';
 
 class GraphicState extends Base {
   render() {
-    const state = `1 0 0 -1 0 ${this.parent.layout.getComputedHeight()} cm`;
+    const stream = `1 0 0 -1 0 ${this.parent.layout.getComputedHeight()} cm`;
 
-    return pdfObject(this.id, pdfStream(state)) + '\n';
+    const graphicState = pdfObject(this.id, pdfStream(stream)) + '\n';
+
+    this.offset = this.root.addOffset(graphicState.length);
+
+    return graphicState;
   }
 }
