+++ packages/react-pdf/src/elements/Rectangle.js
@@ -11,14 +11,18 @@ class Rectangle extends Base {
     const { backgroundColor } = this.style;
     const { left, top, width, height } = this.props;
 
-    const rect = [
+    const stream = [
       '/DeviceRGB cs',
       backgroundColor ? `${toRGB(backgroundColor)} scn` : '',
       backgroundColor ? `${left} ${top} ${width} ${height} re` : '',
       backgroundColor ? 'f' : '',
     ].join('\n');
 
-    return pdfObject(this.id, pdfStream(rect)) + '\n';
+    const rectangle = pdfObject(this.id, pdfStream(stream)) + '\n';
+
+    this.offset = this.root.addOffset(rectangle.length);
+
+    return rectangle;
   }
 }
