+++ packages/react-pdf/src/elements/View.js
@@ -1,9 +1,9 @@
-import PDFEntry from './PDFEntry';
+import Base from './Base';
 import { toRGB } from '../utils/colors';
 import { pdfObject, pdfStream } from '../utils/pdf';
 
-class View extends PDFEntry {
-  render() {
+class View extends Base {
+  async render() {
     const { left, top, width, height } = this.getAbsoluteLayout();
     const { backgroundColor } = this.style;
