+++ packages/react-pdf/src/index.js
@@ -10,7 +10,7 @@ const Page = 'PAGE';
 const Document = 'DOCUMENT';
 
 const pdf = input => {
-  function parse(input) {
+  async function parse(input) {
     const result = input.render();
 
     if (input.props.onRender) {
@@ -20,18 +20,23 @@ const pdf = input => {
     return result;
   }
 
-  function toBlob() {
-    return new Blob([parse(input)], {
+  async function toBlob() {
+    const render = await parse(input);
+
+    return new Blob([render], {
       type: 'application/pdf',
     });
   }
 
-  function toBuffer() {
-    return new Buffer(parse(input));
+  async function toBuffer() {
+    const render = await parse(input);
+    return new Buffer(render);
   }
 
-  function toString() {
-    return parse(input);
+  async function toString() {
+    const render = await parse(input);
+
+    return render;
   }
 
   return {
